+++ ChangeLog.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## UNRELEASED
 
 ## Version 2.13
+- FIX : function getListChildrenOf doesn't take care of last child OF - *28/03/2023* - 2.13.9
 - FIX : Compatibilité v17 > extrafields attribute - *31/01/2023* - 2.13.8
 - FIX : Boucle infinie lors de l'enregistrement d'un OF et de la présence d'un ou plusieurs produits non conformes - Déplacement de code *08/11/2022* - 2.13.7
 - FIX : Compat V16 *10/06/2022* - 2.13.6